+++ static/data/media.js
@@ -1,4 +1,4 @@
-export const TOTAL_MEDIA_CATEGORY = {
+export const CONTENTS_BY_CATEGORY = {
     data: [
         {
             categoryName: "종합/경제",
@@ -65,4 +65,44 @@ export const TOTAL_MEDIA_CATEGORY = {
         }
     ],
     length: 2
+}
+
+export const CONTENTS_BY_MEDIA = {
+    data: [
+        {
+            id: 0,
+            name: "한겨례",
+            iconUrl: "https://s.pstatic.net/static/newsstand/up/2024/0219/nsd145828169.png",
+            editDate: "07월 03일 16:23 편집",
+            imageContent: {
+                imageUrl: "https://www.pressian.com/_resources/10/2024/07/03/2024070311344319469_l.jpg",
+                url: "https://www.pressian.com/pages/articles/2024070311170206107&amp;utm_source=naver&amp;utm_medium=mynews",
+                headline: "봇물처럼 터지는 공공요금 인상…꼭 지금이어야 하나"
+            },
+            contents: [
+                {
+                    url: "",
+                    headline: "봇물처럼 터지는 공공요금 인상…꼭 지금이어야 하나"
+                }
+            ]
+        },
+        {
+            id: 1,
+            name: "세계일보",
+            iconUrl: "https://s.pstatic.net/static/newsstand/2020/logo/light/0604/022.png",
+            editDate: "07월 03일 16:23 편집",
+            imageContent: {
+                imageUrl: "https://it.donga.com/media/__sized__/images/2024/7/3/aafceafdecd94926-thumbnail-1920x1080-70.jpg",
+                url: "https://it.donga.com/105568/",
+                headline: "서울과기대, 실전 같은 '창업캠프'로 대학생 창업 경험 제공해"
+            },
+            contents: [
+                {
+                    url: "",
+                    headline: "봇물처럼 터지는 공공요금 인상…꼭 지금이어야 하나"
+                }
+            ]
+        }
+    ],
+
 }
