+++ utils/events.js
@@ -1,0 +1,5 @@
+export const REMOVE_TOTAL_CATEGORY = "remove-total-category";
+export const removeTotalCategoryEvent = new CustomEvent(REMOVE_TOTAL_CATEGORY);
+
+export const REMOVE_MEDIA_CATEGORY = "remove-media-category";
+export const removeMediaCategoryEvent = new CustomEvent(REMOVE_MEDIA_CATEGORY);
